+++ Assets/Newtonsoft.Json.UnityConverters.Tests/Issues/Issue20_Populating.cs
@@ -1,5 +1,5 @@
-using UnityEngine;
 using NUnit.Framework;
+using UnityEngine;
 
 namespace Newtonsoft.Json.UnityConverters.Tests
 {
@@ -12,15 +12,15 @@ namespace Newtonsoft.Json.UnityConverters.Tests
             // Wrapping in yet another object for Newtonsoft.Json's populate
             // algo to really kick in.
             // Otherwise it just seems to do JSON diffing
-            var input = new MyClass { MyProperty = new Vector3(1, 2, 3) };
-            var json = Serialize(new { MyProperty = new { y = 8 }});
+            var input = new MyClass { myProperty = new Vector3(1, 2, 3) };
+            string json = Serialize(new { myProperty = new { y = 8 } });
             var expectedProp = new Vector3(1, 8, 3);
 
             // Act
             Populate(json, input);
 
             // Assert
-            Assert.AreEqual(expectedProp, input.MyProperty, "JSON: " + json);
+            Assert.AreEqual(expectedProp, input.myProperty, "JSON: " + json);
         }
 
         [Test]
@@ -31,7 +31,7 @@ namespace Newtonsoft.Json.UnityConverters.Tests
             // algo to really kick in.
             // Otherwise it just seems to do JSON diffing
             var input = new MyClass { myField = new Vector3(4, 5, 6) };
-            var json = Serialize(new { myField = new { y = 9 }});
+            string json = Serialize(new { myField = new { y = 9 } });
             var expectedField = new Vector3(4, 9, 6);
 
             // Act
@@ -43,7 +43,7 @@ namespace Newtonsoft.Json.UnityConverters.Tests
 
         private class MyClass
         {
-            public Vector3 MyProperty { get; set; }
+            public Vector3 myProperty { get; set; }
 
             public Vector3 myField;
         }