+++ Packages/Newtonsoft.Json-for-Unity.Converters/Editor/UnityConvertersConfigEditor.cs
@@ -5,6 +5,9 @@ using Newtonsoft.Json.UnityConverters.Configuration;
 using UnityEditor;
 using UnityEditor.AnimatedValues;
 using UnityEngine;
+// Unity 2020 also has a type cache: UnityEditor.TypeCache:
+// https://docs.unity3d.com/2019.2/Documentation/ScriptReference/TypeCache.html
+using TypeCache = Newtonsoft.Json.UnityConverters.Utility.TypeCache;
 
 namespace Newtonsoft.Json.UnityConverters.Editor
 {
@@ -25,9 +28,6 @@ namespace Newtonsoft.Json.UnityConverters.Editor
         private AnimBool _unityConvertersShow;
         private AnimBool _jsonNetConvertersShow;
 
-        private readonly Dictionary<string, Type> _converterTypeByName
-            = new Dictionary<string, Type>();
-
         private GUIStyle _headerStyle;
         private GUIStyle _boldHeaderStyle;
 
@@ -44,7 +44,7 @@ namespace Newtonsoft.Json.UnityConverters.Editor
             try
             {
                 _ = serializedObject;
-            } 
+            }
             catch (Exception)
             {
                 return;
@@ -125,7 +125,7 @@ namespace Newtonsoft.Json.UnityConverters.Editor
         {
             var elements = EnumerateArrayElements(arrayProperty);
             var elementTypes = elements
-                .Select(e => FindType(e.FindPropertyRelative(nameof(ConverterConfig.converterName)).stringValue))
+                .Select(e => TypeCache.FindType(e.FindPropertyRelative(nameof(ConverterConfig.converterName)).stringValue))
                 .ToArray();
             Type[] missingConverters = converterTypes
                 .Where(type => !elementTypes.Contains(type))
@@ -196,7 +196,7 @@ namespace Newtonsoft.Json.UnityConverters.Editor
                 var allConfigsWithType = EnumerateArrayElements(property)
                     .Select(o => (
                         serializedProperty: o,
-                        type: FindType(o.FindPropertyRelative(nameof(ConverterConfig.converterName)).stringValue)
+                        type: TypeCache.FindType(o.FindPropertyRelative(nameof(ConverterConfig.converterName)).stringValue)
                     ));
 
                 foreach (var namespaceGroup in allConfigsWithType.GroupBy(o => GetTypeNamespace(o.type)))
@@ -249,35 +249,6 @@ namespace Newtonsoft.Json.UnityConverters.Editor
             EditorGUILayout.EndFadeGroup();
         }
 
-        private Type FindType(string name)
-        {
-            if (_converterTypeByName.TryGetValue(name, out var type))
-            {
-                return type;
-            }
-            else
-            {
-                // Check this assembly, or if it has AssemblyQualifiedName
-                type = Type.GetType(name);
-                if (type != null)
-                {
-                    _converterTypeByName[name] = type;
-                    return type;
-                }
-
-                // Check all the other assemblies, from last imported to first
-                foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies().Reverse())
-                {
-                    type = assembly.GetType(name);
-                    if (type != null)
-                    {
-                        _converterTypeByName[name] = type;
-                        return type;
-                    }
-                }
-                return null;
-            }
-        }
 
         private static string GetNamespaceHeader(IGrouping<string, (SerializedProperty serializedProperty, Type type)> namespaceGroup)
         {